requirements.txt
matplotlib==3.0.2
numpy==1.16.1
Pillow==5.4.1
prompt-toolkit==2.0.8
ptyprocess==0.6.0
six==1.12.0
torch==1.0.1.post2
torchvision==0.2.1
tqdm==4.31.1
